Modularised & Optimised For SEO
Seperated Admin, Authentication, and Blog routes into seperate blueprints
Added Basic SEO optimisation to homepage and blog pages
Made app more compliant with PEP 8
Improved some aspects of security
