+++ .gitignore
@@ -20,6 +20,7 @@ wheels/
 *.egg-info/
 .installed.cfg
 *.egg
+db_init.py
 
 
 virt/
