+++ blueprints/admin/__init__.py
@@ -1,0 +1,1 @@
+"""Admin Blueprint Containing All Admin Routes"""
