+++ blueprints/admin/static/manage-admins.js
