+++ blueprints/admin/templates/admin/add_article.html
@@ -1,5 +1,5 @@
 {% extends "admin/base.html" %} {% block title %}{% if edit_mode %}Edit{% else
-%}Create{% endif %} Post{% endblock %} {% block head %}
+%}Create{% endif %} Article{% endblock %} {% block head %}
 <script
   src="https://cdn.tiny.cloud/1/ult3ksjejkbf2yzhxkktmab1xmlnkylkaq0ws88hrbf4i76i/tinymce/7/tinymce.min.js"
   referrerpolicy="origin"
@@ -7,7 +7,7 @@
 {% endblock %} {% block body %}
 <div class="container mx-auto px-4 py-12">
   <div class="max-w-4xl mx-auto bg-gray-900 rounded-xl shadow-lg p-8">
-    <form id="blog-post-form" method="POST" enctype="multipart/form-data">
+    <form id="add-article-form" method="POST" enctype="multipart/form-data">
       {{ form.hidden_tag() }}
       <meta name="csrf-token" content="{{ csrf_token() }}" />
 
@@ -94,7 +94,7 @@
           type="submit"
           class="bg-blue-600 hover:bg-blue-700 text-white px-8 py-3 rounded-full transition-all font-semibold hover:cursor-pointer"
         >
-          {% if edit_mode %}Update Post{% else %}Create Post{% endif %}
+          {% if edit_mode %}Update Article{% else %}Create Article{% endif %}
         </button>
       </div>
     </form>
@@ -126,7 +126,7 @@
 
     automatic_uploads: true,
     // For image uploads
-    oolbar: "image",
+
     images_upload_handler: (blobInfo, progress) => {
       return new Promise((resolve, reject) => {
         const formData = new FormData();
@@ -226,7 +226,9 @@
             hiddenInput.type = "hidden";
             hiddenInput.name = "thumbnail_link";
             hiddenInput.id = "thumbnailLink";
-            document.getElementById("blog-post-form").appendChild(hiddenInput);
+            document
+              .getElementById("add-article-form")
+              .appendChild(hiddenInput);
           }
           hiddenInput.value = json.link;
         } else {
