+++ blueprints/admin/templates/admin/dashboard.html
@@ -1,6 +1,6 @@
 {% extends "admin/base.html" %}
 
-{% block title %}Admin Panel -  Welcome{% endblock %}
+{% block title %}Welcome{% endblock %}
 
 {% block body %}
         <!-- Stats Cards -->
@@ -8,7 +8,7 @@
           <div class="bg-gray-900 rounded-lg p-6 border border-gray-800">
             <div class="flex items-center justify-between">
               <div>
-                <p class="text-gray-400 text-sm">Total Posts</p>
+                <p class="text-gray-400 text-sm">Total Articles</p>
                 <p class="text-2xl font-bold text-white">
                   {{ total_articles }}
                 </p>
@@ -37,33 +37,32 @@
           <h3 class="text-lg font-semibold text-white mb-4">Quick Actions</h3>
           <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
             
-              <a href="{{ url_for('add_post') }}"
+              <a href="{{ url_for('admin.add_article') }}"
                 class="bg-blue-600 hover:bg-blue-700 text-white px-6 py-4 rounded-lg transition-colors duration-200 flex items-center space-x-3 hover:cursor-pointer"
               >
                 <i class="fas fa-plus"></i>
-                <span>Create New Post</span>
+                <span>Create New Article</span>
           </a>
             
             <a
-            href="{{ url_for('manage_admins') }}#add-admin"
+            href="{{ url_for('admin.manage_admins') }}#add-admin"
               class="bg-green-600 hover:bg-green-700 text-white px-6 py-4 rounded-lg transition-colors duration-200 flex items-center space-x-3 hover:cursor-pointer"
             >
               <i class="fas fa-user-plus"></i>
               <span>Add New Admin</span>
         </a>
             <a
-            href="{{ url_for('logout') }}"
+            href="{{ url_for('auth.logout') }}"
               class="bg-red-600 hover:bg-red-700 text-white px-6 py-4 rounded-lg transition-colors duration-200 flex items-center space-x-3 hover:cursor-pointer"
             >
               <i class="fas fa-right-from-bracket"></i>
               <span>Logout</span>
-      </a>
+                </a>
           </div>
         </div>
-      </div>
+      
     </div>
+    
 
   
 {% endblock %}
-  </body>
-</html>
