+++ blueprints/admin/templates/admin/sidebar.html
@@ -13,7 +13,7 @@
           Dashboard
         </h2>
         <a
-          href="{{ url_for('admin_dashboard') }}"
+          href="{{ url_for('admin.admin_dashboard') }}"
           class="flex items-center px-4 py-3 text-gray-300 hover:bg-gray-800 hover:text-white rounded-lg transition-colors duration-200"
         >
           <i class="fas fa-tachometer-alt mr-3"></i>
@@ -30,16 +30,16 @@
         </h2>
         <div class="space-y-1">
           <a
-            href="{{ url_for('manage_posts') }}"
+            href="{{ url_for('admin.manage_articles') }}"
             class="flex items-center px-4 py-3 text-gray-300 hover:bg-gray-800 hover:text-white rounded-lg transition-colors duration-200"
           >
             <i class="fas fa-edit mr-3"></i>
-            <span>Manage Posts</span>
+            <span>Manage Articles</span>
           </a>
 
           {% if current_user.is_superadmin %}
           <a
-            href="{{ url_for('manage_admins') }}"
+            href="{{ url_for('admin.manage_admins') }}"
             class="flex items-center px-4 py-3 text-gray-300 hover:bg-gray-800 hover:text-white rounded-lg transition-colors duration-200"
           >
             <i class="fas fa-users-cog mr-3"></i>
@@ -57,7 +57,7 @@
           Account
         </h2>
         <a
-          href="/logout"
+          href="{{ url_for('auth.logout') }}"
           class="flex items-center px-4 py-3 text-gray-300 hover:bg-red-800 hover:text-white rounded-lg transition-colors duration-200"
         >
           <i class="fas fa-sign-out-alt mr-3"></i>
