+++ blueprints/auth/__init__.py
@@ -1,0 +1,1 @@
+"""Auth Blueprint Containing All Authentication Routes"""
