+++ blueprints/auth/templates/auth/reset_password.html
@@ -8,51 +8,42 @@
       href="{{ url_for('static', filename='css/main.css') }}"
     />
     <script src="{{ url_for('static', filename='js/alert-dismissal.js') }}"></script>
-    <title>Login</title>
+    <title>Reset Password</title>
   </head>
   <body class="bg-[#0f0f0f] text-white">
     {% include 'alert.html' %}
     <div
       class="flex p-8 items-center md:w-[560px] md:h-[520px] min-w-[320px] h-[510px] absolute inset-1/2 transform -translate-x-1/2 -translate-y-1/2 flex-col bg-[#1a1a1a] rounded-xl shadow-md border border-[#2e2e2e] backdrop-blur-sm bg-opacity-90 justify-center"
     >
-      <h1 class="text-2xl font-bold mb-2 text-white">Admin Panel</h1>
-      <h2 class="text-lg text-gray-300 mb-6">Please sign in to continue</h2>
+      <h1 class="text-2xl font-bold mb-2 text-white">Reset Password</h1>
+      <h2 class="text-lg text-gray-300 mb-6">Please type your new password</h2>
 
-      <form method="POST" action="{{ url_for('login') }}" class="w-full">
+      <form
+        method="POST"
+        action="{{ url_for('auth.reset_password', token=token) }}"
+        class="w-full"
+      >
         {{ form.hidden_tag() }}
 
         <div class="mb-4">
-          <label for="username" class="block text-gray-300 mb-1"
-            >Username</label
+          <label for="new_password" class="block text-gray-300 mb-1"
+            >New Password</label
           >
-          {{ form.username(class_="w-full px-4 py-2 border border-gray-600
+          {{ form.new_password(class_="w-full px-4 py-2 border border-gray-600
           bg-[#111111] text-white rounded-lg focus:outline-none focus:ring-2
-          focus:ring-blue-600") }} {% for error in form.username.errors %}
+          focus:ring-blue-600") }} {% for error in form.new_password.errors %}
           <span class="text-red-500 text-sm">{{ error }}</span>
           {% endfor %}
-          <p class="mt-2 text-sm text-blue-400 underline">
-            <a
-              href="{{ url_for('forgot_username') }}"
-              class="hover:text-blue-300"
-              >Forgot Username?</a
-            >
-          </p>
         </div>
 
         <div class="mb-4">
-          {{ form.password.label(class_="block text-gray-300 mb-1") }} {{
-          form.password(class_="w-full px-4 py-2 border border-gray-600
-          bg-[#111111] text-white rounded-lg focus:outline-none focus:ring-2
-          focus:ring-blue-600") }} {% for error in form.password.errors %}
+          {{ form.confirm_password.label(class_="block text-gray-300 mb-1") }}
+          {{ form.confirm_password(class_="w-full px-4 py-2 border
+          border-gray-600 bg-[#111111] text-white rounded-lg focus:outline-none
+          focus:ring-2 focus:ring-blue-600") }} {% for error in
+          form.confirm_password.errors %}
           <span class="text-red-500 text-sm">{{ error }}</span>
           {% endfor %}
-          <p class="mt-2 text-sm text-blue-400 underline">
-            <a
-              href="{{ url_for('forgot_password') }}"
-              class="hover:text-blue-300"
-              >Forgot Password?</a
-            >
-          </p>
         </div>
 
         <div>
