+++ blueprints/blog/__init__.py
@@ -1,0 +1,1 @@
+"""Blog Blueprint Containing All Public Blog Routes"""
