+++ blueprints/blog/templates/blog/view_post.html
@@ -1,13 +1,107 @@
 <!DOCTYPE html>
 <html lang="en" class="bg-white">
   <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
+    <!-- SEO Meta Tags -->
+    <title>{{ article.title }} - Dr. Fassih Ul Abbas | Medical Blog</title>
+    <meta
+      name="description"
+      content="{{ article.description[:155] if article.description else (article.content | striptags)[:155] }}..."
+    />
+    <meta
+      name="keywords"
+      content="Dr. Fassih Ul Abbas, medical blog, {{ article.tags.replace(',', ', ') if article.tags else 'healthcare, medicine' }}"
+    />
+    <meta name="author" content="Dr. Fassih Ul Abbas" />
+    <meta name="robots" content="index, follow" />
+    <link rel="canonical" href="{{ request.url }}" />
+
+    <!-- Open Graph / Facebook -->
+    <meta property="og:type" content="article" />
+    <meta
+      property="og:title"
+      content="{{ article.title }} - Dr. Fassih Ul Abbas"
+    />
+    <meta
+      property="og:description"
+      content="{{ article.description[:155] if article.description else (article.content | striptags)[:155] }}..."
+    />
+    <meta
+      property="og:image"
+      content="{{ article.thumbnail or 'https://picsum.photos/1200/630?random=' ~ article.id }}"
+    />
+    <meta property="og:url" content="{{ request.url }}" />
+    <meta property="og:site_name" content="Dr. Fassih Ul Abbas Medical Blog" />
+    <meta property="article:author" content="Dr. Fassih Ul Abbas" />
+    <meta
+      property="article:published_time"
+      content="{{ article.created_at.isoformat() }}"
+    />
+    {% if article.updated_at %}
+    <meta
+      property="article:modified_time"
+      content="{{ article.updated_at.isoformat() }}"
+    />
+    {% endif %} {% if article.tags %} {% for tag in article.tags.split(',') %}
+    <meta property="article:tag" content="{{ tag.strip() }}" />
+    {% endfor %} {% endif %}
+
+    <!-- Twitter Card -->
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta
+      name="twitter:title"
+      content="{{ article.title }} - Dr. Fassih Ul Abbas"
+    />
+    <meta
+      name="twitter:description"
+      content="{{ article.description[:155] if article.description else (article.content | striptags)[:155] }}..."
+    />
+    <meta
+      name="twitter:image"
+      content="{{ article.thumbnail or 'https://picsum.photos/1200/630?random=' ~ article.id }}"
+    />
+
+    <!-- Schema.org JSON-LD -->
+    <script type="application/ld+json">
+      {
+        "@context": "https://schema.org",
+        "@type": "Article",
+        "headline": "{{ article.title }}",
+        "description": "{{ article.description if article.description else (article.content | striptags)[:155] }}",
+        "image": "{{ article.thumbnail or 'https://picsum.photos/1200/630?random=' ~ article.id }}",
+        "datePublished": "{{ article.created_at.isoformat() }}",
+        {% if article.updated_at %}
+        "dateModified": "{{ article.updated_at.isoformat() }}",
+        {% endif %}
+        "author": {
+          "@type": "Person",
+          "name": "Dr. Fassih Ul Abbas",
+          "jobTitle": "Medical Professional",
+          "url": "{{ url_for('index', _external=True) }}"
+        },
+        "publisher": {
+          "@type": "Organization",
+          "name": "Dr. Fassih Ul Abbas Medical Blog",
+          "url": "{{ url_for('index', _external=True) }}"
+        },
+        "mainEntityOfPage": {
+          "@type": "WebPage",
+          "@id": "{{ request.url }}"
+        },
+        "url": "{{ request.url }}"
+        {% if article.tags %},
+        "keywords": "{{ article.tags.replace(',', ', ') }}"
+        {% endif %}
+      }
+    </script>
+
+    <!-- Stylesheets -->
     <link
       href="{{ url_for('static', filename='css/main.css') }}"
       rel="stylesheet"
     />
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>{{ article.title }} - Dr. Fassih Ul Abbas</title>
     <link
       rel="stylesheet"
       href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"
@@ -16,9 +110,12 @@
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
     />
-    <meta
-      name="description"
-      content="{{ article.description[:150] if article.description else article.content[:150] }}..."
+
+    <!-- Preload critical resources -->
+    <link
+      rel="preload"
+      href="{{ article.thumbnail or 'https://picsum.photos/1920/1080?random=' ~ article.id }}"
+      as="image"
     />
   </head>
 
@@ -62,7 +159,7 @@
             <div class="flex gap-2">
               <!-- Edit Button -->
               <a
-                href="{{ url_for('edit_post', post_id=article.id) }}"
+                href="{{ url_for('admin.edit_article', article_id=article.id) }}"
                 class="flex items-center bg-blue-100 hover:bg-blue-200 text-blue-800 px-4 py-2 rounded-lg transition-colors"
               >
                 <span class="material-symbols-outlined mr-1 text-sm">edit</span>
@@ -223,7 +320,7 @@
             <form
               id="deleteForm"
               method="POST"
-              action="{{ url_for('delete_article', article_id=article.id) }}"
+              action="{{ url_for('admin.delete_article', article_id=article.id) }}"
             >
               <button
                 type="submit"
