+++ templates/alert.html
@@ -1,7 +1,7 @@
 {% with messages = get_flashed_messages() %} {% if messages %} {% for msg in
 messages %}
 <div
-  class="z-20 flex items-center p-4 text-red-800 rounded-lg dark:text-red-400 alert border border-primary absolute top-0 w-full animate-slide-in"
+  class="z-20 flex items-center p-4 text-red-800 rounded-lg dark:text-red-400 alert border border-primary absolute top-0 w-full animate-slide-in {% block alert_color %}{% endblock %}"
   role="alert"
 >
   <svg
@@ -15,7 +15,7 @@ messages %}
       d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3H8a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z"
     />
   </svg>
-  <div class="ms-3 text-sm font-medium">{{msg|safe}}</div>
+  <div class="ms-3 text-sm font-medium">{{msg}}</div>
   <button
     type="button"
     class="banner-close-btn ms-auto -mx-1.5 -my-1.5 bg-red-50 text-red-500 rounded-lg focus:ring-2 focus:ring-red-400 p-1.5 hover:bg-red-200 inline-flex items-center justify-center h-8 w-8 dark:bg-gray-800 dark:text-red-400 dark:hover:bg-gray-700"
