+++ templates/sidebar.html
@@ -72,7 +72,7 @@
           </li>
           <li>
             <a
-              href="{{ url_for('all_posts') }}"
+              href="{{ url_for('blog.all_articles') }}"
               class="flex items-center px-4 py-3 text-gray-700 hover:text-blue-600 hover:bg-blue-50 rounded-lg transition duration-200 group"
             >
               <span
